FIX: GLS wird sinnvoll gelöst.

+++ src/steifigkeitsmatrix.f90
@@ -24,8 +24,8 @@ subroutine StabsteifigkeitLokalKIIK(Stabs, zuBefuellendeMatrix)
   use kenngroessen
   implicit none
   type(stab) :: Stabs
-  real, dimension(3,3) :: zuBefuellendeMatrix
-  real :: l
+  real*8, dimension(3,3) :: zuBefuellendeMatrix
+  real*8 :: l
 
   l=sqrt((koordinatenmatrix(Stabs%K2,1) - koordinatenmatrix(Stabs%K1,1))**2  &
         +(koordinatenmatrix(Stabs%K2,2) - koordinatenmatrix(Stabs%K1,2))**2)   !Stablängen bestimmen "Wurzel((x2-x1)^2-(y2-y1)^2)"
@@ -44,8 +44,8 @@ subroutine StabsteifigkeitLokalIKKI(Stabs, zuBefuellendeMatrix)
   use kenngroessen
   implicit none
   type(stab) :: Stabs
-  real, dimension(3,3) :: zuBefuellendeMatrix
-  real :: l
+  real*8, dimension(3,3) :: zuBefuellendeMatrix
+  real*8 :: l
 
   l=sqrt((koordinatenmatrix(Stabs%K2,1) - koordinatenmatrix(Stabs%K1,1))**2  &
         +(koordinatenmatrix(Stabs%K2,2) - koordinatenmatrix(Stabs%K1,2))**2)   !Stablängen bestimmen "Wurzel((x2-x1)^2-(y2-y1)^2)"
@@ -64,8 +64,8 @@ subroutine StabsteifigkeitLokalIKIK(Stabs, zuBefuellendeMatrix)
   use kenngroessen
   implicit none
   type(stab) :: Stabs
-  real, dimension(3,3) :: zuBefuellendeMatrix
-  real :: l
+  real*8, dimension(3,3) :: zuBefuellendeMatrix
+  real*8 :: l
 
   l=sqrt((koordinatenmatrix(Stabs%K2,1) - koordinatenmatrix(Stabs%K1,1))**2  &
         +(koordinatenmatrix(Stabs%K2,2) - koordinatenmatrix(Stabs%K1,2))**2)   !Stablängen bestimmen "Wurzel((x2-x1)^2-(y2-y1)^2)"
